feat(rendering-compat): R2473 counter value type i32→i64 全链路 widening (基建)
CSS 计数器值规范上无界，旧 parse_counter_action / CounterActionValue.value /
painter.counters / get_counter / format_counter_text / counter_style_body /
format_counter_alpha/roman 全 i32 → CJK counter 测试用 10^16 量级 counter-reset
（simp/trad-chinese 等）静默溢出丢弃声明。

widening 8 site 全 i64（覆盖 ~9.2×10^18），Fixed(first) cast i64 适配。
unit test 加 i64 大值断言（n 9999999999999 / n 10000000000000000 → Some，原 i32 溢出 None）。

net-0 self-source（无 CJK algorithm 前），spec-correct 基建，unlock 后续 CJK
formal/informal（6+ flip 候选）须配合各 ~100 行 multiplicative-additive 算法。

旁证 dead-end：css-content quotes A/B（paint_quotes 跳过）= 零 delta（double-render
假说证伪，已 revert）；css-cascade 实测 26/49=53.1%（stale 快照 32.7% 已订正）。

门禁全绿：make test 13145/0/74，clippy/fmt clean，
product-smoke welcome 17.03% struct 全 PASS，product-smoke-legacy 51/51 struct PASS。

+++ crates/css-parser/src/values/parse_extended.rs
@@ -404,7 +404,10 @@ pub struct CounterActionValue {
     /// 计数器名称。
     pub name: String,
     /// 增量或重置值，None 表示默认（increment=1, reset=0）。
-    pub value: Option<i32>,
+    ///
+    /// R2473：i64（原 i32）—— CSS 计数器值在规范上无界，CJK counter 测试用 10^16 量级
+    /// counter-reset（simp/trad-chinese 等），i32 会静默溢出丢弃声明。i64 覆盖到 ~9.2×10^18。
+    pub value: Option<i64>,
 }
 
 /// 解析单个计数器操作值。
@@ -422,7 +425,7 @@ pub fn parse_counter_action(input: &str) -> Option<CounterActionValue> {
         return None;
     }
     let value = if parts.len() > 1 {
-        Some(parts[1].parse::<i32>().ok()?)
+        Some(parts[1].parse::<i64>().ok()?)
     } else {
         None
     };
@@ -445,8 +448,8 @@ pub fn parse_counter_list(input: &str) -> Option<Vec<CounterActionValue>> {
             return None;
         }
         // 检查下一个 token 是否为整数
-        let value = if tokens.peek().is_some_and(|t| t.parse::<i32>().is_ok()) {
-            tokens.next().and_then(|t| t.parse::<i32>().ok())
+        let value = if tokens.peek().is_some_and(|t| t.parse::<i64>().is_ok()) {
+            tokens.next().and_then(|t| t.parse::<i64>().ok())
         } else {
             None
         };
